+++ birdy/client/utils.py
@@ -120,78 +120,3 @@ def from_owslib(value, data_type):
         # todo: boundingbox
         pass
     return value
-
-
-def is_notebook():
-    """Return whether or not this function is executed in a notebook environment."""
-    try:
-        from IPython import get_ipython
-    except ImportError:
-        return False
-
-    try:
-        shell = get_ipython().__class__.__name__
-        if shell == 'ZMQInteractiveShell':
-            return True   # Jupyter notebook or qtconsole
-        elif shell == 'TerminalInteractiveShell':
-            return False  # Terminal running IPython
-        else:
-            return False  # Other type (?)
-    except NameError:
-        return False      # Probably standard Python interpreter
-
-
-def input2widget(inpt):
-    """Return a Notebook widget to enter values for the input."""
-    import ipywidgets as ipyw
-
-    if not isinstance(inpt, Input):
-        raise ValueError
-
-    typ = inpt.dataType
-    opt = inpt.allowedValues
-
-    # String default
-    default = inpt.defaultValue
-
-    # Object default
-    odefault = from_owslib(inpt.defaultValue, inpt.dataType)
-
-    kwds = dict(description=inpt.title)
-    if opt:
-        vopt = [from_owslib(o, typ) for o in opt]
-        if inpt.maxOccurs == 1:
-            if len(opt) < 3:
-                out = ipyw.RadioButtons(options=vopt, **kwds)
-            else:
-                out = ipyw.Dropdown(options=vopt, **kwds)
-        else:
-            out = ipyw.SelectMultiple(options=vopt, value=[inpt.defaultValue], description=inpt.title)
-    elif typ.endswith('float'):
-        out = ipyw.FloatText(value=odefault, **kwds)
-    elif typ.endswith('boolean'):
-        out = ipyw.Checkbox(value=odefault, **kwds)
-    elif typ.endswith('integer'):
-        out = ipyw.IntText(value=odefault, **kwds)
-    elif typ.endswith('positiveInteger'):
-        out = ipyw.BoundedIntText(value=odefault, min=1E-16, **kwds)
-    elif typ.endswith('nonNegativeInteger'):
-        out = ipyw.BoundedIntText(value=odefault, min=0, **kwds)
-    elif typ.endswith('string'):
-        out = ipyw.Text(value=odefault, placeholder=inpt.abstract, **kwds)
-    elif typ.endswith('anyURI'):
-        out = ipyw.Text(value=odefault, placeholder=inpt.abstract, **kwds)
-    elif typ.endswith('time'):
-        out = ipyw.Text(value=default, placeholder='YYYY-MM-DD', **kwds)
-    elif typ.endswith('date'):
-        out = ipyw.Text(value=default, placeholder='hh-mm-ss', **kwds)
-    elif typ.endswith('dateTime'):
-        out = ipyw.Text(value=default, placeholder='YYYY-MM-DDThh-mm-ss', **kwds)
-    elif typ.endswith('angle'):
-        out = ipyw.BoundedFloatText(min=0, max=360, **kwds)
-    elif typ.endswith('ComplexData'):
-        out = ipyw.Text(description=inpt.title)
-    else:
-        raise AttributeError("Data type not recognized {}".format(typ))
-
-    return out
