+++ tests/test_dependencies.py
@@ -1,0 +1,3 @@
+def test_dependencies():
+    from birdy.dependencies import ipywidgets as widgets
+    from birdy.dependencies import IPython
